Restructure Python SDK Part 2
