+++ examples/requirements.txt
@@ -1,4 +1,4 @@
--e ../../
+-e ../
 httpx
 typing_extensions
 python-dotenv
